+++ pr_tlb_miss_second_iteration_offset_K40/pr_tlb_miss_second_iteration_offset_K40.cu
@@ -38,7 +38,7 @@ __device__ void P_chasing0(int mark, int *A, int iterations, int *B, int *C, lon
 }
 
 //////////min page size 4kb = 4096b = 32 * 128.
-__device__ void P_chasing1(int mark, int *A, int iterations, int *B, int *C, long long int *D, int starting_index, float clock_rate, int data_stride){	
+__device__ void P_chasing1(int mark, int *A, long long int iterations, int *B, int *C, long long int *D, int starting_index, float clock_rate, int data_stride){	
 	
 	int j = starting_index;/////make them in the same page, and miss near in cache lines
 	
@@ -127,7 +127,7 @@ __device__ void P_chasing2(int mark, int *A, long long int iterations, int *B, i
 
 __global__ void tlb_latency_test(int *A, long long int iterations, int *B, int *C, long long int *D, float clock_rate, long long int mod, int data_stride){
 	
-	int reduced_iter = iterations;
+	long long int reduced_iter = iterations;
 	if(reduced_iter > 512){
 		reduced_iter = 512;
 	}else if(reduced_iter < 16){
@@ -204,7 +204,7 @@ int main(int argc, char **argv)
 		CPU_data_in = (int*)malloc(sizeof(int) * data_size);
 		init_cpu_data(CPU_data_in, mod, data_stride, mod);
 		
-		int reduced_iter = iterations;
+		long long int reduced_iter = iterations;
 		if(reduced_iter > 512){
 			reduced_iter = 512;
 		}else if(reduced_iter < 16){
